Refactors components and adds responsive container
Refactors structure of how home page is rendered with helper function to
generate responsive wrapper.
