+++ .eslintrc.json
@@ -14,8 +14,9 @@
           "MemberExpression": 1
       }],
       "object-curly-spacing": ["error", "always"],
-      "react/jsx-uses-react": "error",   
-      "react/jsx-uses-vars": "error" 
+      "react/jsx-uses-react": "error",
+      "react/jsx-uses-vars": "error" ,
+      "eol-last": 2
   },
   "plugins": ["react"]
 }
