+++ package.json
@@ -8,8 +8,8 @@
     "eslint-plugin-react": "^7.12.4",
     "history": "^4.7.2",
     "markdown-it": "^8.4.2",
-    "markdown-it-music": "github:music-markdown/markdown-it-music",
     "node-sass": "^4.11.0",
+    "npm": "^6.8.0",
     "query-string": "^6.2.0",
     "react": "^16.8.1",
     "react-bootstrap": "^1.0.0-beta.5",
