+++ src/index.js
@@ -1,13 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components/App';
 import registerServiceWorker from './registerServiceWorker';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import App from './components/App.js';
 
 ReactDOM.render(
-  <Router>
-    <Route path='/' exact component={App} />
-  </Router>,
-  document.getElementById('root'));
+  <App />,
+  document.getElementById('root')
+);
 registerServiceWorker();
